requirements.txt
numpy==2.4.3
pandas==3.0.1
plotly==6.6.0
psycopg2_binary==2.9.11
pytest==9.0.2
scikit_learn==1.8.0
SQLAlchemy==2.0.48
streamlit==1.54.0
torch==2.9.1+cu130
yfinance==1.1.0
